+++ CHANGELOG.md
@@ -7,6 +7,28 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+### Testing
+
+-   Created test for the sign up attempt using duplicate email
+
+## [0.28.1] - 2022-12-13
+
+### Fixes
+
+-   Trailing slash is added to URL when going back to "Sign In" from "Forgot Password" for consistency (fixes https://github.com/supertokens/supertokens-auth-react/issues/625)
+
+## [0.28.0] - 2022-12-13
+
+### Breaking change
+
+-   Changed styles for social login buttons
+
+## [0.27.4] - 2022-12-06
+
+### Testing
+
+-   Updated test to match new behaviour of the website SDK - not rethrowing refresh errors during `doesSessionExist`
+
 ## [0.27.3] - 2022-11-25
 
 ### Adds